MachineListening UGens adapted for Windows
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@7088 a380766d-ff14-0410-b294-a243070f3f08

+++ Source/plugins/BeatTrack.cpp
@@ -783,7 +783,7 @@ void finaldecision(BeatTrack *unit) {
 	float phasebeforeamort= ((float)unit->m_bestphase/unit->m_torround);
 	
 	//add phase to compensate for ELAPSEDTIME
-	unit->m_currphase= unit->m_phase = fmod(phasebeforeamort+phaseelapsed,1.0);
+	unit->m_currphase= unit->m_phase = fmod(phasebeforeamort+phaseelapsed,(float)1.0);
 	
 }
